Fix link macro parsing and change orphan checking
Fixed the link macro parsing, which didn't do exactly as I wanted, due to a
link in the form <<link "Text" $somevariable>> would give the link text, rather
than noting that there is no passage to parse. Had to use a regex lookahead
function in order to get it to work, but it seems to flawlessly handle all link
types I throw at it. So I'll call it complete.

The second thing I did is change the way I do orphan checking due to a bug in
the way I did it that caused the wrong passages to get ignored and in general
was highly inconsistent, due to my misconception that deleting from a link
while iterating caused side effects. The way it now is handled is it keeps
track of which passages are linked to a passage, and then running a count to
find passages with no passages linking to them. This also gets me reverse link
checking which I can write to the errorfile as well.

With braid itself, I changed around some of the parameters, and used my new
found knowledge of how to get arguments to take no options to add flags for
broken links and to show the reverse passage links. These are -b and -r
respectively, -e is now used to specify the file that any of the errors will be
written to, but at the moment that is just the broken links/orphans, and the
reverse passage list.