Enable config file support for braid
Braid is now able to load a configuration file that is used to set it's
variables and start the game building. It must be located within the current
working directory when braid is executed, and it must be named either braid.cfg
or braid.ini. An example is located at doc/braid.cfg.example.