Add functionality to skip stripping newlines
As someone requested, in order to make debugging of javascript and such within
a story possible, added an option to prevent braid from stripping the newlines
from passages. By default it will strip them, but a tag name can be given to
skip only for those tags. There is also the special option all, which will tell
braid to not strip newlines on all passages. Also passages tagged script have
special consideration, getting an additional newline before and after the
passage, so the passage is all on its own lines.