Add image base64 encoding and embedding to passages marked Twine.media
Added the capability for braid to automatically load image files listed in
passage text and converting them to a base64 encoded string that can be
embedded within the html and the browser to load. Requires the image files to
be relative to the current working directory of braid execution. Supported
filetypes are png, gif, and jpg.